@remotion/effects: Fix scale() effect to use subpath export
- Move scale implementation to src/scale/index.ts for proper subpath export
- Create src/scale.ts as re-export file (matching blur/wave pattern)
- Remove scale from main index.ts exports
- Add src/scale.ts to bundle.ts entrypoints
- Fix import path in scale implementation to use parent directory
- Remove root-level scale.ts file (replaced by proper subpath structure)
