--- packages/effects/scale.ts
@@ -1,1 +1,0 @@
-export {scale, type ScaleParams} from './src/scale.js';
