+++ packages/effects/src/index.ts
@@ -1,4 +1,3 @@
 export {EffectInternals} from './effect-internals.js';
 export {halftoneSchema, type HalftoneParams} from './halftone.js';
 export {tintSchema, type TintParams} from './tint.js';
-export {scale, type ScaleParams} from './scale.js';
